+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.6.0
+* NEW: Added force selection setting for latest available date
 * NEW: Added force selection setting for current period
 * FIX: Scroll bars don't display correctly